--- src/main/java/dev/nincodedo/ninbot/common/CommandNameEnum.java
@@ -1,10 +1,0 @@
-package dev.nincodedo.ninbot.common;
-
-public interface CommandNameEnum {
-
-    String name();
-
-    default String get() {
-        return this.name().toLowerCase();
-    }
-}
