+++ src/main/java/dev/nincodedo/ninbot/common/SlashCommandRegistration.java
@@ -1,6 +1,6 @@
 package dev.nincodedo.ninbot.common;
 
-import dev.nincodedo.ninbot.components.command.SlashCommand;
+import dev.nincodedo.ninbot.common.command.SlashCommand;
 import lombok.extern.slf4j.Slf4j;
 import net.dv8tion.jda.api.events.ReadyEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
