+++ src/main/java/dev/nincodedo/ninbot/common/command/CommandNameEnum.java
@@ -1,4 +1,4 @@
-package dev.nincodedo.ninbot.components.common;
+package dev.nincodedo.ninbot.common.command;
 
 public interface CommandNameEnum {
