+++ src/main/java/dev/nincodedo/ninbot/components/countdown/CountdownCommandName.java
@@ -1,0 +1,7 @@
+package dev.nincodedo.ninbot.components.countdown;
+
+import dev.nincodedo.ninbot.common.command.CommandNameEnum;
+
+public enum CountdownCommandName implements CommandNameEnum {
+    LIST, CREATE
+}
