+++ src/main/java/dev/nincodedo/ninbot/components/event/EventCommandName.java
@@ -1,6 +1,6 @@
 package dev.nincodedo.ninbot.components.event;
 
-import dev.nincodedo.ninbot.common.CommandNameEnum;
+import dev.nincodedo.ninbot.common.command.CommandNameEnum;
 
 public enum EventCommandName implements CommandNameEnum {
     LIST, PLAN
