+++ src/main/java/dev/nincodedo/ninbot/components/fun/eightball/Magic8BallCommand.java
@@ -1,6 +1,6 @@
 package dev.nincodedo.ninbot.components.fun.eightball;
 
-import dev.nincodedo.ninbot.components.command.SlashCommand;
+import dev.nincodedo.ninbot.common.command.SlashCommand;
 import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
 import net.dv8tion.jda.api.interactions.commands.build.OptionData;
