+++ src/main/java/dev/nincodedo/ninbot/components/fun/hugemoji/HugemojiCommand.java
@@ -1,7 +1,7 @@
 package dev.nincodedo.ninbot.components.fun.hugemoji;
 
 import dev.nincodedo.ninbot.common.message.MessageAction;
-import dev.nincodedo.ninbot.components.command.AbstractCommand;
+import dev.nincodedo.ninbot.common.command.AbstractCommand;
 import lombok.extern.slf4j.Slf4j;
 import net.dv8tion.jda.api.events.message.priv.PrivateMessageReceivedEvent;
