+++ src/main/java/dev/nincodedo/ninbot/components/poll/PollRepository.java
@@ -10,5 +10,5 @@ public interface PollRepository extends CrudRepository<Poll, Long> {
 
     Optional<Poll> findByMessageIdAndPollOpen(String messageId, boolean isPollOpen);
 
-    Optional<Poll> findById(Long id);
+    List<Poll> findByIds(List<Long> ids);
 }
