+++ src/main/java/dev/nincodedo/ninbot/components/tally/TallyCommand.java
@@ -1,6 +1,6 @@
 package dev.nincodedo.ninbot.components.tally;
 
-import dev.nincodedo.ninbot.components.command.SlashCommand;
+import dev.nincodedo.ninbot.common.command.SlashCommand;
 import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
 import net.dv8tion.jda.api.interactions.commands.OptionMapping;
 import net.dv8tion.jda.api.interactions.commands.OptionType;
