+++ src/main/java/dev/nincodedo/ninbot/components/tally/TallyCommandName.java
@@ -1,6 +1,6 @@
 package dev.nincodedo.ninbot.components.tally;
 
-import dev.nincodedo.ninbot.common.CommandNameEnum;
+import dev.nincodedo.ninbot.common.command.CommandNameEnum;
 
 public enum TallyCommandName implements CommandNameEnum {
     ADD, GET, SUBTRACT
